+++ RELEASE_NOTES.md
@@ -28,6 +28,7 @@ Enhancements:
  * mapfishapp: mapfishapp packages were renamed as org.georchestra.mapfishapp, see  https://github.com/georchestra/georchestra/pull/69
  * downloadform: the packages were renamed as org.georchestra.dlform, see  https://github.com/georchestra/georchestra/pull/72
  * extractorapp: packges were renamed as org.georchestra.extractorapp, see https://github.com/georchestra/georchestra/pull/71 
+ * security-proxy: the module's packages were renamed as org.georchestra.security, see  https://github.com/georchestra/georchestra/pull/73
 
 Bug fixes:
  * mapfishapp: fixed erroneous WMSC2WMS mapping, which prevented printing of the GeoBretagne OSM baselayer, see https://github.com/georchestra/georchestra/commit/159bd4f24ecb21b9c76f76d27c1736ec1040f0ab